requirements.txt
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
colorama==0.3.9
coverage==4.4.1
Django==1.11.2
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.3
docutils==0.14
Faker==0.8.6
future==0.16.0
h2o==3.10.4.8
idna==2.6
ipaddress==1.0.18
mongoengine==0.9.0
numpy==1.13.1
pandas==0.20.3
pexpect==4.0.1
ptyprocess==0.5
pycparser==2.17
pymongo==2.8.1
python-dateutil==2.6.1
pytz==2017.2
pyxdg==0.25
PyYAML==3.11
requests==2.18.4
selenium==3.5.0
six==1.10.0
statistics==1.0.3.5
tabulate==0.7.7
tbselenium==0.2.1
text-unidecode==1.0
urllib3==1.22
xlrd==1.0.0
